% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{create_analysis_obj}
\alias{create_analysis_obj}
\title{create_analysis_obj}
\usage{
create_analysis_obj(datasets, datasetpath)

create_analysis_obj(datasets, datasetpath)
}
\arguments{
\item{datasetpath}{path to dataset}

\item{sces}{SingleCellExperiment objects}
}
\value{
list of lists, each representing a dataset with the attributes data, params and config

list of lists, each representing a dataset with the attributes data, params and config
}
\description{
maps dataset name, data and params, reads dataset specific analysis params

maps dataset name, data and params, reads dataset specific analysis params
}
