% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana.R
\name{get_liana_agg}
\alias{get_liana_agg}
\title{get_liana_agg}
\usage{
get_liana_agg(res, dataset, new = FALSE)
}
\arguments{
\item{res}{result from run_liana_wrap}

\item{dataset}{element of 'analysis' object}

\item{new}{\if{html}{\out{<boolean>}} if true (re)run aggregation else read previous results in}
}
\description{
get_liana_agg
}
