% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{get_subsets}
\alias{get_subsets}
\title{Get Subsets}
\usage{
get_subsets(dataset_name, subset_category)
}
\arguments{
\item{dataset_name}{name of dataset}

\item{subset_category}{category of subset which is a metadata column}
}
\value{
subsets
}
\description{
Get Subsets
}
