% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsets.R
\name{get_subsets_of_dataset}
\alias{get_subsets_of_dataset}
\title{get_subsets_of_dataset To add subset names via analysis_params$other$<dataset_name>$subs$<subset_category>}
\usage{
get_subsets_of_dataset(datasets, datasetname, metacol)
}
\arguments{
\item{datasetname}{name of dataset}

\item{metacol}{for example "seurat_clusters"}
}
\description{
get_subsets_of_dataset To add subset names via analysis_params$other$<dataset_name>$subs$<subset_category>
}
