% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{loadPackage}
\alias{loadPackage}
\title{basic}
\usage{
loadPackage(package, fromGithub = F, fromBioconductor = F)
}
\arguments{
\item{package}{the name of the package}

\item{fromGithub}{if the package shall be installed from Github}

\item{fromBioconductor}{if the package shall be installed from Bioconductor}
}
\value{
package is loaded
}
\description{
a package is loaded or installed and loaded
}
\examples{
loadPackage("base")
loadPackage("ying14/yingtools2", 1)
}
\author{
Thalla
}
\keyword{install}
\keyword{load}
\keyword{package}
