% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(dataset_names, datapath, analysis_params, func)
}
\arguments{
\item{dataset_names}{names of datasets}

\item{datapath}{path to data in mounted folder, assumption1: data folder is called 'data', assumption2: 'original_name' contains the full name of the file}

\item{analysis_params}{params}

\item{func}{reader function}
}
\value{
datasets (list)
}
\description{
read_data
}
\author{
Thalla
}
