% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ds_analysis_params}
\alias{read_ds_analysis_params}
\title{Read Dataset Analysis Params}
\usage{
read_ds_analysis_params(dataset_name, datasetpath)
}
\arguments{
\item{datasetpath}{path of dataset}

\item{subset_name}{name of subset}
}
\description{
read in dataset specific parameters
}
