% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_liana_methods}
\alias{read_liana_methods}
\title{read_liana_methods}
\usage{
read_liana_methods(dataset, clustercolname)
}
\arguments{
\item{ds_name}{dataset name}
}
\description{
same as 'res'/result from liana_wrap
}
