% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{rename_all}
\alias{rename_all}
\title{rename_all uses analysis params to rename metadata columns}
\usage{
rename_all(sces, analysis)
}
\arguments{
\item{sces}{a list of singleCellExperiment objects}

\item{analysis}{a list of additional information for each dataset in sces (analysis_params and paths)}
}
\description{
for each dataset the metadata columns are renamed according to the information in the analysis_params
}
