% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana.R
\name{run_liana_plots}
\alias{run_liana_plots}
\title{run_liana_plots}
\usage{
run_liana_plots(agg, analysis$all, "seurat_clusters", lte = c(0.01, 0.05))
}
\arguments{
\item{agg}{aggregated results from LIANA}

\item{dataset}{element of 'analysis', has data, config, ...}

\item{cluster_col_name}{name of clustering column that was used with LIANA so far, if a list is provided loops over list}

\item{ltes}{filter aggregate_rank according to lower than or equals the given value, if a list is provided loops over list}

\item{ntops}{number of top n results chosen for dotplot, if a list is provided loops over list}
}
\description{
plot aggregation results and save the plots, loops over ltes, ntops, lig_rec
}
\author{
Thalla
}
