% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana.R
\name{run_liana_wrap}
\alias{run_liana_wrap}
\title{run_liana_wrap}
\usage{
run_liana_wrap(dataset, cluster_col_name, new)
}
\arguments{
\item{dataset}{element of list 'analysis'}

\item{cluster_col_name}{name of the column that defines the cell clusters}

\item{new}{(boolean), if true (re)run liana_wrap}
}
\description{
run_liana_wrap
}
