% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{save_tool_config}
\alias{save_tool_config}
\title{save_tool_config}
\usage{
save_tool_config(datasets, datasetpath, toolname)

save_tool_config(datasets, datasetpath, toolname)
}
\arguments{
\item{datasets}{list of datasets}

\item{toolname}{name of tool}
}
\value{
config table

config table
}
\description{
Creates and saves a config file for a specific tool. Each dataset is added with the relevant paths. File is saved in the version folder.

Creates and saves a config file for a specific tool. Each dataset is added with the relevant paths. File is saved in the version folder.
}
\author{
Thalla
}
\keyword{config}
\keyword{init}
