% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana.R
\name{src_trgt_heatmap}
\alias{src_trgt_heatmap}
\title{src_trgt_heatmap}
\usage{
src_trgt_heatmap(agg_plot, title, nrow, ncol, figpath)
}
\arguments{
\item{agg_plot}{aggregation data, eventually preprocessed for plotting}

\item{title}{title of plot}

\item{nrow}{for filterint table matrix: table(source, target)\link{1:nrow, 1:ncol}}

\item{ncol}{see nrow}
}
\description{
Heatmap showing the number of significant interactions between clusters
}
\author{
Thalla
}
